# [executor] parallelize execute and commit in benchmark