--- a/execution/executor-benchmark/benches/executor_benchmark.rs
+++ b/execution/executor-benchmark/benches/executor_benchmark.rs
@@ -5,6 +5,7 @@ use criterion::{criterion_group, criterion_main, measurement::Measurement, Batch
 use executor_benchmark::{
     create_storage_service_and_executor, TransactionExecutor, TransactionGenerator,
 };
+use std::sync::Arc;
 
 pub const NUM_ACCOUNTS: usize = 1000;
 pub const SMALL_BLOCK_SIZE: usize = 500;
@@ -21,10 +22,12 @@ fn executor_benchmark<M: Measurement + 'static>(c: &mut Criterion<M>) {
 
     let (_db, executor) = create_storage_service_and_executor(&config);
     let parent_block_id = executor.committed_block_id();
+    let executor = Arc::new(executor);
 
     let mut generator = TransactionGenerator::new(genesis_key, NUM_ACCOUNTS);
+    let (commit_tx, _commit_rx) = std::sync::mpsc::channel();
 
-    let mut executor = TransactionExecutor::new(executor, parent_block_id);
+    let mut executor = TransactionExecutor::new(executor, parent_block_id, commit_tx);
     let txns = generator.gen_account_creations(SMALL_BLOCK_SIZE);
     for txn_block in txns {
         executor.execute_block(txn_block);