@msmouse 2021-09-14
Any change in TPS that we observed?