@zekun000 2021-09-14
before: Version: 30000. execute time: 540 ms. commit time: 148 ms. TPS: 1452. Accumulative TPS: 1415
after: Version: 30000. latency: 859 ms, execute time: 701 ms. commit time: 157 ms. TPS: 1425. Accumulative TPS: 1563

i guess not much improvement